+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+### Added
+-  Python iGraph lib support
 ### Fixed
 - Google docstring return style
 - Typos in docstrings